packages/backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
aiohttp==3.10.5
# psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.9.2
redis==5.0.8
